Let's just put it all in... ...and try to remember what I changed and fixed: - Dialogs like the cd copy dialog are not closed after the copy is finished anymore. I alwayw   found that annoying. - Added some more error handling. - Added a whatsthis hint for the DVS sizes (many users are confused by the 4.4 instead of 4.7) - enhanced debugging output (devices) - K3b now tries to determine the max writing speed for audio on-the-fly writing by decoding small   parts of every source. - Some internal changes.
svn path=/trunk/kdeextragear-1/k3b/; revision=380067

+++ libk3b/projects/datadvd/k3bdvdjob.cpp
@@ -500,7 +500,7 @@ void K3bDvdJob::slotVerificationFinished( bool success )
 
 void K3bDvdJob::cleanup()
 {
-  if( d->imageError || m_canceled || m_doc->removeImages() ) {
+  if( !m_doc->onTheFly() && ( d->imageError || m_canceled || m_doc->removeImages() ) ) {
     if( QFile::exists( m_doc->tempDir() ) ) {
       QFile::remove( m_doc->tempDir() );
       emit infoMessage( i18n("Removed image file %1").arg(m_doc->tempDir()), K3bJob::SUCCESS );
